third_party/WebKit/LayoutTests/http/tests/inspector/network/network-filters.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>

async function test()
{
    await InspectorTest.clearNetworkCache();

    InspectorTest.recordNetwork();

    InspectorTest.makeFetch("resources/style.css", {}, ensureAllResources);
    InspectorTest.makeFetch("resources/abe.png", {}, () => {
      // Ensures result is cached.
      InspectorTest.makeFetch("resources/abe.png", {}, ensureAllResources);
      ensureAllResources();
    });
    InspectorTest.makeFetch("missing/foo.bar", {}, ensureAllResources);

    var filterChecks = [{
        filterText: "-.css",
        isRegex: false
    }, {
        filterText: "-.png",
        isRegex: false
    }, {
        filterText: "css",
        isRegex: false
    }, {
        filterText: "",
        isRegex: false
    }, {
        filterText: ".*",
        isRegex: true
    }, {
        filterText: ".*\\..*",
        isRegex: true
    }, {
        filterText: ".*\\.png",
        isRegex: true
    }, {
        filterText: "NOTHINGTOMATCH",
        isRegex: true
    }, {
        filterText: "",
        isRegex: true
    }, {
        filterText: "png",
        isRegex: false
    }, {
        filterText: "-missing",
        isRegex: false
    }, {
        filterText: "is:from-cache",
        isRegex: false
    }, {
        filterText: "-is:from-cache",
        isRegex: false
    }];

    var resourceCount = 0;
    var totalResourceCount = 4;
    function ensureAllResources()
    {
        if (++resourceCount >= totalResourceCount)
            checkFilters();
    }

    function checkFilters()
    {
        for (var filterObj of filterChecks) {
            InspectorTest.addResult("filterText: " + filterObj.filterText);
            InspectorTest.addResult("isRegex: " + filterObj.isRegex);

            setNetworkLogFilter(filterObj.filterText, filterObj.isRegex);

            var nodes = UI.panels.network._networkLogView.flatNodesList();
            var foundNodesCount = 0;
            for (var i = 0; i < nodes.length; i++) {
                if (!nodes[i][Network.NetworkLogView._isFilteredOutSymbol])
                    foundNodesCount++;
            }

            InspectorTest.addResult("Found results: " + foundNodesCount);
            InspectorTest.addResult("");
        }
        InspectorTest.completeTest();
    }

    /**
     * @param {string} value
     * @param {RegExp} isRegex
     */
    function setNetworkLogFilter(value, isRegex)
    {
        UI.panels.network._networkLogView._textFilterUI._regexCheckBox.checked = isRegex;
        UI.panels.network._networkLogView._textFilterUI.setValue(value);
        UI.panels.network._networkLogView._filterChanged(null); // event not used in this method, so passing null
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests fetch network filters</p>
</body>
</html>
