third_party/WebKit/LayoutTests/platform/mac-mac10.10/http/tests/inspector/network/network-filters-expected.txt

Tests fetch network filters

filterText: -.css
isRegex: false
Found results: 3

filterText: -.png
isRegex: false
Found results: 2

filterText: css
isRegex: false
Found results: 1

filterText: 
isRegex: false
Found results: 4

filterText: .*
isRegex: true
Found results: 4

filterText: .*\..*
isRegex: true
Found results: 4

filterText: .*\.png
isRegex: true
Found results: 2

filterText: NOTHINGTOMATCH
isRegex: true
Found results: 0

filterText: 
isRegex: true
Found results: 4

filterText: png
isRegex: false
Found results: 2

filterText: -missing
isRegex: false
Found results: 3

filterText: is:from-cache
isRegex: false
Found results: 2

filterText: -is:from-cache
isRegex: false
Found results: 2
